LSP: Add devkit cmake tests
Validate devkit's cmake compilation.

+++ LSP/A-Devkit/Host/compile.rb
@@ -28,21 +28,20 @@ def run
   return if run_and_check_cmd(download_cmd, equip, /DOWNLOADPASSED/)
   toolchain_filename = Dir.entries("#{Dir.home()}/#{toolchains_url.sub('http://','')}/#{arch}")[-1]
   toolchain_dirname = File.basename(toolchain_filename, ".tar.xz")
-  if equip.send_sudo_cmd(["tar xvf ~/#{toolchains_url.sub('http://','')}/#{arch}/#{toolchain_filename} -C /opt", "echo 'INSTALLATIONPASSED'"], /INSTALLATIONPASSED/, 120)
-    set_result(FrameworkConstants::Result[:fail], "Error trying to install toolchain")
-    return
+  if !Dir.exists?("/opt/#{toolchain_dirname}/bin")
+    if equip.send_sudo_cmd(["tar xvf ~/#{toolchains_url.sub('http://','')}/#{arch}/#{toolchain_filename} -C /opt", "echo 'INSTALLATIONPASSED'"], /INSTALLATIONPASSED/, 120)
+      set_result(FrameworkConstants::Result[:fail], "Error trying to install toolchain")
+      return
+    end
   end
   compile_test = File.join(File.dirname(__FILE__), 'compile.sh')
   toolchain_path = "/opt/#{toolchain_dirname}/bin"
   devkit_url =  @test_params.var_devkit_url
-  c_app_path = File.join(File.dirname(__FILE__), '..', 'hello.c')
-  cpp_app_path = File.join(File.dirname(__FILE__), '..', 'hello.cpp')
-  pthread_app_path = File.join(File.dirname(__FILE__), '..', 'thread-ex.c')
-
   test_cmd ="#{compile_test} #{toolchain_path} #{devkit_url} " \
             "#{File.join(File.dirname(__FILE__), '..', 'hello.c')} " \
             "#{File.join(File.dirname(__FILE__), '..', 'hello.cpp')} " \
-            "#{File.join(File.dirname(__FILE__), '..', 'thread-ex.c')}"
+            "#{File.join(File.dirname(__FILE__), '..', 'thread-ex.c')} " \
+            "#{File.join(File.dirname(__FILE__), '..', 'cmake-sample')}"
   return if run_and_check_cmd(test_cmd, equip, /All host-side devkit checks passed/, 900)
   set_result(FrameworkConstants::Result[:pass], "Host-side devkit passed all tests")
 end
